# docs/openapi.yaml
openapi: 3.1.0
info:
  title: SkyMP Backend API
  version: 2.0.0
  description: Canonical API v2 for launchers, dashboards, and SkyMP game servers.
servers:
  - url: http://localhost:4000
...
paths:
...
  /api/v2/launcher/servers/{key}:
    parameters:
      - $ref: '#/components/parameters/ServerKey'
    get:
      tags: [Launcher]
      operationId: getLauncherServer
      parameters:
        - $ref: '#/components/parameters/SessionHeader'
      responses:
        '200': { $ref: '#/components/responses/JsonSuccess' }
        '404': { $ref: '#/components/responses/JsonError' }
  /api/v2/launcher/manifest:
    get:
      tags: [Launcher]
      operationId: getLauncherManifest
      responses:
        '200': { $ref: '#/components/responses/Collection' }
        '503': { $ref: '#/components/responses/JsonError' }
...
  /api/v2/launcher/client/download:
    get:
      tags: [Launcher]
      operationId: downloadClient
      responses:
        '200':
          description: Client archive
          content:
            application/zip:
              schema: { type: string, contentEncoding: binary }
        '404': { $ref: '#/components/responses/JsonError' }
...
  /api/v2/game-servers/{key}/heartbeat:
    post:
      tags: [Game servers]
      operationId: recordHeartbeat
      parameters: [{ $ref: '#/components/parameters/ServerKey' }]
      requestBody: { $ref: '#/components/requestBodies/JsonBody' }
      responses:
        '200': { $ref: '#/components/responses/JsonSuccess' }
        '403': { $ref: '#/components/responses/JsonError' }
  /api/v2/game-servers/{key}/sessions/{token}:
    get:
      tags: [Game servers]
      operationId: getGameSession
      parameters:
        - { $ref: '#/components/parameters/ServerKey' }
        - { name: token, in: path, required: true, schema: { type: string } }
      responses:
        '200': { $ref: '#/components/responses/JsonSuccess' }
        '404': { $ref: '#/components/responses/JsonError' }
  /api/v2/game-servers/{key}/profiles/{id}/access:
    get:
      tags: [Game servers]
      operationId: checkProfileAccess
      parameters:
        - { $ref: '#/components/parameters/ServerKey' }
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        '200': { $ref: '#/components/responses/JsonSuccess' }
  /api/v2/game-servers/{key}/profiles/{id}/factions:
    post:
      tags: [Game servers]
      operationId: assignProfileFaction
      parameters:
        - { $ref: '#/components/parameters/ServerKey' }
        - { name: id, in: path, required: true, schema: { type: integer } }
        - { $ref: '#/components/parameters/AuthTokenHeader' }
      requestBody: { $ref: '#/components/requestBodies/JsonBody' }
      responses:
        '201': { $ref: '#/components/responses/JsonSuccess' }
  /api/v2/game-servers/{key}/profiles/{id}/factions/{assignmentId}:
    delete:
      tags: [Game servers]
      operationId: removeProfileFaction
      parameters:
        - { $ref: '#/components/parameters/ServerKey' }
        - { name: id, in: path, required: true, schema: { type: integer } }
        - { name: assignmentId, in: path, required: true, schema: { type: string } }
        - { $ref: '#/components/parameters/AuthTokenHeader' }
      responses:
        '200': { $ref: '#/components/responses/JsonSuccess' }
  /api/v2/game-servers/{key}/sessions/{token}/balance:
    get:
      tags: [Game servers]
      operationId: getSessionBalance
      parameters:
        - { $ref: '#/components/parameters/ServerKey' }
        - { name: token, in: path, required: true, schema: { type: string } }
      responses:
        '200': { $ref: '#/components/responses/JsonSuccess' }
  /api/v2/game-servers/{key}/sessions/{token}/purchases:
    post:
      tags: [Game servers]
      operationId: createSessionPurchase
      parameters:
        - { $ref: '#/components/parameters/ServerKey' }
        - { name: token, in: path, required: true, schema: { type: string } }
        - { $ref: '#/components/parameters/AuthTokenHeader' }
      requestBody: { $ref: '#/components/requestBodies/JsonBody' }
      responses:
        '200': { $ref: '#/components/responses/JsonSuccess' }
...
components:
  securitySchemes:
    dashboardBearer:
      type: http
      scheme: bearer
  parameters:
    ServerKey:
      name: key
      in: path
      required: true
      schema: { type: string }
...
    SessionHeader:
      name: X-Session
      in: header
      required: false
      schema: { type: string }
    AuthTokenHeader:
      name: X-Auth-Token
      in: header
      required: true
      schema: { type: string }
  requestBodies:
    JsonBody:
      required: true
      content:
        application/json:
          schema: { type: object, additionalProperties: true }
  responses:
    JsonSuccess:
      description: Successful JSON response
      content:
        application/json:
          schema: { type: object, additionalProperties: true }
    Collection:
      description: Collection response
      content:
        application/json:
          schema:
            type: object
            required: [items, total]
            properties:
              items: { type: array, items: {} }
              total: { type: integer, minimum: 0 }
    JsonError:
      description: Structured API v2 error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details: {}